pyproject.toml
[project]
name = "property-notifier"
version = "0.1.0"
description = "A property alert notification microservice."
authors = [
    {name = "FrenyCS",email = "53234343+FrenyCS@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "celery[redis] (>=5.4.0,<6.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)"
]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pylint = "^3.3.6"
isort = "^6.0.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
